pyproject.toml
[project]
name = "opencode-orchestrator"
version = "0.1.0"
description = "Multi-model orchestrator-worker extension for OpenCode"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.25.0",
    "requests>=2.32.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
]

[project.scripts]
opencode-orchestrator = "opencode_orchestrator:main"
opencode-orchestrator-ui = "opencode_orchestrator.backend.app:start"
opencode-orchestrate = "opencode_orchestrator.cli:cli_orchestrate"
opencode-orchestrate-status = "opencode_orchestrator.cli:cli_status"
opencode-orchestrate-config = "opencode_orchestrator.cli:cli_config"
opencode-orchestrate-worker = "opencode_orchestrator.cli:cli_configure_worker"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opencode_orchestrator"]

[tool.hatch.build.targets.wheel.shared-data]
"frontend/dist" = "opencode_orchestrator/frontend_dist"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "httpx>=0.28.0"]
